# code/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorDetail:
    properties:
      code:
        example: invalid_input
        type: string
      error:
        example: invalid request body
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handlers.ErrorDetail'
    type: object
...
  transaction.ChangeTransactionInput:
    properties:
      account_code:
        type: string
      account_transfer:
        type: string
      amount:
        type: integer
      category_code:
        type: string
      date:
        type: string
      description:
        type: string
      is_transfer:
        type: boolean
    type: object
  transaction.CreateSingleTransaction:
    properties:
      account_code:
        type: string
      account_transfer:
        type: string
      amount:
        type: integer
      category_code:
        type: string
      date:
        type: string
      description:
        type: string
      is_transfer:
        type: boolean
    required:
    - account_code
    - amount
    - category_code
    - date
    - description
    - is_transfer
    type: object
  transaction.CreateTransactionInput:
    properties:
      transactions:
        items:
          $ref: '#/definitions/transaction.CreateSingleTransaction'
        type: array
    required:
    - transactions
    type: object
  transaction.CreateTransactionResponse:
    properties:
      transactions:
        items:
          $ref: '#/definitions/transaction.TransactionResponseItem'
        type: array
    type: object
  transaction.GetTransactionResponse:
    properties:
      transaction:
        $ref: '#/definitions/transaction.TransactionResponseItem'
    type: object
  transaction.PaginationInfo:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  transaction.TransactionResponse:
    properties:
      pagination:
        $ref: '#/definitions/transaction.PaginationInfo'
      transactions:
        items:
          $ref: '#/definitions/transaction.TransactionResponseItem'
        type: array
    type: object
  transaction.TransactionResponseItem:
    properties:
      account_code:
        type: string
      account_transfer:
        type: string
      amount:
        type: integer
      category_code:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      transfer_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for managing accounts, categories, transactions, reports, and suggestions.
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /transactions:
    get:
      description: Returns filtered transactions with pagination metadata
      parameters:
      - description: Page size
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Sort field, prepend '-' for descending
        enum:
        - DATE
        - AMOUNT
        - UPDATED
        - -DATE
        - -AMOUNT
        - -UPDATED
        in: query
        name: sort
        type: string
      - description: Comma-separated account codes
        in: query
        name: account_code
        type: string
      - description: Comma-separated category codes
        in: query
        name: category_code
        type: string
      - description: Space- or comma-separated description terms, supports exclusions
          with '-'
        in: query
        name: description
        type: string
      - description: Comma-separated operations
        enum:
        - REVENUE
        - EXPENSE
        - TRANSFER
        in: query
        name: operation
        type: string
      - description: Minimum absolute amount
        in: query
        name: min_amount
        type: integer
      - description: Maximum absolute amount
        in: query
        name: max_amount
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transaction.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Creates one or more transactions for the authenticated user
      parameters:
      - description: Transactions payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transaction.CreateTransactionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/transaction.CreateTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create transactions
      tags:
      - transactions
  /transactions/{id}:
    delete:
      description: Deletes a transaction
      parameters:
      - description: Transaction ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - transactions
    get:
      description: Returns a transaction by its ID
      parameters:
      - description: Transaction ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transaction.GetTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - transactions
    patch:
      consumes:
      - application/json
      description: Updates an existing transaction
      parameters:
      - description: Transaction ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Transaction changes
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transaction.ChangeTransactionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transaction.GetTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update transaction
      tags:
      - transactions
securityDefinitions:
  BearerAuth:
    description: JWT token using the format "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"